requirements.txt
requests
edge-tts
gtts
python-dotenv
